Also clean up the octopus container

+++ scripts/linux/wait-octopus-with-db-restart.sh
@@ -18,14 +18,16 @@ until $(curl --max-time 5 --output /dev/null --silent --head --fail http://local
     sleep 60
 
     # The database sometimes fails us, so do a hard reset of the database container
-    systemctl stop docker-octopusdeploy
-    systemctl stop docker-mssql
-    docker stop mssql
-    docker rm mssql
-    docker stop octopusdeploy
-    docker rm octopusdeploy
-    systemctl start docker-mssql
-    systemctl start docker-octopusdeploy
+    if [[ ! $(curl --max-time 5 --output /dev/null --silent --head --fail http://localhost) ]]; then
+      systemctl stop docker-octopusdeploy
+      systemctl stop docker-mssql
+      docker stop mssql
+      docker rm mssql
+      docker stop octopusdeploy
+      docker rm octopusdeploy
+      systemctl start docker-mssql
+      systemctl start docker-octopusdeploy
+    fi
 done
 
 exit 0
